+++ app/src/main/java/se/miun/dt170/antonsskafferi/ui/kitchen/KitchenBongContainerView.java
@@ -1,4 +1,4 @@
-package se.miun.dt170.antonsskafferi.ui.bong;
+package se.miun.dt170.antonsskafferi.ui.kitchen;
 
 import android.content.Context;
 import android.view.LayoutInflater;
@@ -7,12 +7,12 @@ import androidx.constraintlayout.widget.ConstraintLayout;
 
 import se.miun.dt170.antonsskafferi.R;
 
-public class BongKitchenListView extends ConstraintLayout {
-    public BongKitchenListView(Context context) {
+public class KitchenBongContainerView extends ConstraintLayout {
+    public KitchenBongContainerView(Context context) {
         super(context);
 
         LayoutInflater inflater = (LayoutInflater) context.getSystemService(Context.LAYOUT_INFLATER_SERVICE);
-        inflater.inflate(R.layout.bong_kitchen_list, this, true);
+        inflater.inflate(R.layout.kitchen_bong_container_view, this, true);
 
     }
 }
