+++ app/src/main/res/layout/activity_kitchen.xml
@@ -8,4 +8,10 @@
     android:orientation="horizontal"
     tools:context=".activity.KitchenActivity">
 
+    <include
+        layout="@layout/kitchen_bong_container_view"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:visibility="gone"
+        tools:visibility="visible" />
 </LinearLayout>
