+++ app/src/main/res/layout/kitchen_bong_header_view.xml
