+++ app/src/main/res/values/attrs_bong_kitchen_list_view.xml
@@ -1,5 +1,5 @@
 <resources>
-    <declare-styleable name="BongKitchenListView">
+    <declare-styleable name="KitchenBongContainerView">
         <attr name="exampleString" format="string" />
         <attr name="exampleDimension" format="dimension" />
         <attr name="exampleColor" format="color" />
